implementing controller and service

+++ src/main/java/br/com/banco/entities/Account.java
@@ -2,6 +2,7 @@ package br.com.banco.entities;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 import javax.persistence.*;
 
 @Entity
@@ -13,7 +14,7 @@ public class Account {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_conta")
-    private Long idConta;
+    private Integer idConta;
 
     @Column(name = "nome_responsavel", nullable = false)
     private String nomeResponsável;
